Use pytest instead of unittest
