+++ .travis.yml
@@ -15,7 +15,7 @@ install:
 - pip install -r requirements.txt
 
 script:
-- python manage.py test --failfast -v 2 --settings=tests.test_settings
+- pytest tests
 
 deploy:
   - provider: codedeploy
