+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+# addopts = --nomigrations --reuse-db
+DJANGO_SETTINGS_MODULE = tests.test_settings
