+++ requirements.txt
@@ -27,3 +27,4 @@ datefinder==0.6.0
 raven
 django-debug-toolbar
 pysolr==3.6.0
+pytest-django==3.3.3
